Extract playlist scrollbar
